Optimize SQL query using left join
This is for displaying and linking to the tickets that have the
specified tag. If a tag appears on 5 tickets, the count will be 5. The
previous solution was quite inefficient, but this new one with a manual
SQL query is much more so. This affects the tags#index action and the
respective models and views.